Separate range types to their own source files.
