+++ src/range_from_exclusive.rs
@@ -1,0 +1,81 @@
+use core::ops::{
+    Bound::{self, Excluded, Unbounded},
+    RangeBounds,
+};
+
+/// A range only bounded exclusively below.
+///
+/// The `RangeFromExclusive` contains all values with `x > start`.
+///
+/// *Note*: Overflow in the [`Iterator`] implementation (when the contained data type reaches its
+/// numerical limit) is allowed to panic, wrap, or saturate. This behavior is defined by the
+/// implementation of the [`Step`] trait. For primitive integers, this follows the normal rules, and
+/// respects the overflow checks profile (panic in debug, wrap in release). Note also that overflow
+/// happens earlier than you might assume: the overflow happens in the call to next that yields the
+/// maximum value, as the range must be set to a state to yield the next value.
+///
+/// # Example
+/// `RangeFromExclusive`s can be created directly, as follows:
+///
+/// ```
+/// use more_ranges::RangeFromExclusive;
+///
+/// let range = RangeFromExclusive {
+///     start: 1,
+/// };
+/// ```
+///
+/// [`Iterator`]: core::iter::Iterator
+/// [`Step`]: core::iter::Step
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub struct RangeFromExclusive<Idx> {
+    /// The lower bound of the range (exclusive).
+    pub start: Idx,
+}
+
+impl<T> RangeBounds<T> for RangeFromExclusive<T> {
+    #[inline]
+    fn start_bound(&self) -> Bound<&T> {
+        Excluded(&self.start)
+    }
+    #[inline]
+    fn end_bound(&self) -> Bound<&T> {
+        Unbounded
+    }
+}
+
+impl<'a, T> RangeBounds<T> for RangeFromExclusive<&'a T> {
+    #[inline]
+    fn start_bound(&self) -> Bound<&T> {
+        Excluded(self.start)
+    }
+    #[inline]
+    fn end_bound(&self) -> Bound<&T> {
+        Unbounded
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::RangeFromExclusive;
+    use core::ops::{
+        Bound::{Excluded, Unbounded},
+        RangeBounds,
+    };
+
+    #[test]
+    fn range_from_exclusive_range_bounds() {
+        let range = RangeFromExclusive { start: 1 };
+
+        assert_matches!(range.start_bound(), Excluded(1));
+        assert_matches!(range.end_bound(), Unbounded);
+    }
+
+    #[test]
+    fn range_from_exclusive_range_bounds_borrowed() {
+        let range = RangeFromExclusive { start: &1 };
+
+        assert_matches!(RangeBounds::<usize>::start_bound(&range), Excluded(1));
+        assert_matches!(RangeBounds::<usize>::end_bound(&range), Unbounded);
+    }
+}
