+++ src/range_from_exclusive_to_exclusive.rs
@@ -1,0 +1,79 @@
+use core::ops::{
+    Bound::{self, Excluded},
+    RangeBounds,
+};
+
+/// A range bounded exclusively below and above.
+///
+/// The `RangeFromExclusiveToExclusive` contains all values with `x > start` and x < end`. It is
+/// empty unless `start < end + 1`.
+///
+/// # Example
+/// `RangeFromExclusiveToExclusive`s can be created directly, as follows:
+///
+/// ```
+/// use more_ranges::RangeFromExclusiveToExclusive;
+///
+/// let range = RangeFromExclusiveToExclusive {
+///     start: 1,
+///     end: 4,
+/// };
+/// ```
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub struct RangeFromExclusiveToExclusive<Idx> {
+    /// The lower bound of the range (exclusive).
+    pub start: Idx,
+    /// The upper bound of the range (exclusive).
+    pub end: Idx,
+}
+
+impl<T> RangeBounds<T> for RangeFromExclusiveToExclusive<T> {
+    #[inline]
+    #[must_use]
+    fn start_bound(&self) -> Bound<&T> {
+        Excluded(&self.start)
+    }
+    #[inline]
+    #[must_use]
+    fn end_bound(&self) -> Bound<&T> {
+        Excluded(&self.end)
+    }
+}
+
+impl<'a, T> RangeBounds<T> for RangeFromExclusiveToExclusive<&'a T> {
+    #[inline]
+    #[must_use]
+    fn start_bound(&self) -> Bound<&T> {
+        Excluded(self.start)
+    }
+    #[inline]
+    #[must_use]
+    fn end_bound(&self) -> Bound<&T> {
+        Excluded(self.end)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::RangeFromExclusiveToExclusive;
+    use core::ops::{
+        Bound::Excluded,
+        RangeBounds,
+    };
+
+    #[test]
+    fn range_from_exclusive_to_exclusive_range_bounds() {
+        let range = RangeFromExclusiveToExclusive { start: 1, end: 3 };
+
+        assert_matches!(range.start_bound(), Excluded(1));
+        assert_matches!(range.end_bound(), Excluded(3));
+    }
+
+    #[test]
+    fn range_from_exclusive_to_exclusive_range_bounds_borrowed() {
+        let range = RangeFromExclusiveToExclusive { start: &1, end: &3 };
+
+        assert_matches!(RangeBounds::<usize>::start_bound(&range), Excluded(1));
+        assert_matches!(RangeBounds::<usize>::end_bound(&range), Excluded(3));
+    }
+}
