+++ src/range_from_exclusive_to_inclusive.rs
@@ -1,0 +1,79 @@
+use core::ops::{
+    Bound::{self, Excluded, Included},
+    RangeBounds,
+};
+
+/// A range bounded exclusively below and inclusively above.
+///
+/// The `RangeFromExclusiveToInclusive` contains all values with `x > start` and `x <= end`. It is
+/// empty unless `start < end`.
+///
+/// # Example
+/// `RangeFromExclusiveToInclusive`s can be created directly, as follows:
+///
+/// ```
+/// use more_ranges::RangeFromExclusiveToInclusive;
+///
+/// let range = RangeFromExclusiveToInclusive {
+///     start: 1,
+///     end: 4,
+/// };
+/// ```
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub struct RangeFromExclusiveToInclusive<Idx> {
+    /// The lower bound of the range (exclusive).
+    pub start: Idx,
+    /// The upper bound of the range (inclusive).
+    pub end: Idx,
+}
+
+impl<T> RangeBounds<T> for RangeFromExclusiveToInclusive<T> {
+    #[inline]
+    #[must_use]
+    fn start_bound(&self) -> Bound<&T> {
+        Excluded(&self.start)
+    }
+    #[inline]
+    #[must_use]
+    fn end_bound(&self) -> Bound<&T> {
+        Included(&self.end)
+    }
+}
+
+impl<'a, T> RangeBounds<T> for RangeFromExclusiveToInclusive<&'a T> {
+    #[inline]
+    #[must_use]
+    fn start_bound(&self) -> Bound<&T> {
+        Excluded(self.start)
+    }
+    #[inline]
+    #[must_use]
+    fn end_bound(&self) -> Bound<&T> {
+        Included(self.end)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::RangeFromExclusiveToInclusive;
+    use core::ops::{
+        Bound::{Excluded, Included},
+        RangeBounds,
+    };
+
+    #[test]
+    fn range_from_exclusive_to_inclusive_range_bounds() {
+        let range = RangeFromExclusiveToInclusive { start: 1, end: 3 };
+
+        assert_matches!(range.start_bound(), Excluded(1));
+        assert_matches!(range.end_bound(), Included(3));
+    }
+
+    #[test]
+    fn range_from_exclusive_to_inclusive_range_bounds_borrowed() {
+        let range = RangeFromExclusiveToInclusive { start: &1, end: &3 };
+
+        assert_matches!(RangeBounds::<usize>::start_bound(&range), Excluded(1));
+        assert_matches!(RangeBounds::<usize>::end_bound(&range), Included(3));
+    }
+}
